+++ libsock/dlp.c
@@ -52,7 +52,7 @@ char * dlp_errorlist[] = {
 char * dlp_strerror(int error) {
   if (error<0)
     error = -error;
-  if (error>sizeof(dlp_errorlist))
+  if ((unsigned int)error>sizeof(dlp_errorlist))
     return "Unknown error";
   else
     return dlp_errorlist[error];
